fix(actor/edit): picture correct size (fix #107)
Previously, if the profile picture was larger than the canvas, it would
overflow during the library loading. This issue was caused because the
picture was shown before the library at replaced it. Switching to
`height: 100%; display: none` fixed this.

If the picture was smaller than the canvas, it would occasionally be
scaled down to 100px. This issue happened because if the image was small
enough, the library would load before the modal had finished loading,
giving a parent size of 0. Updating the event from `show.bs.modal` to
`shown.bs.modal` fixed this.

Signed-off-by: sblablaha <sblablaha@gmail.com>